src/requirements.txt
numpy==1.13.1
PyYAML==5.1
scipy==0.19.1
spglib==1.13.0.post6
